fix sampling for best pipeline
